+++ Availability/Sources/ModuleAvailability/CoreMotionAvailability.swift
